% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lettergen.R
\name{lettergen}
\alias{lettergen}
\title{A Random Letter Generator Function}
\usage{
lettergen(n)
}
\arguments{
\item{n}{Number of letters generated}
}
\description{
This function allows you to generate a random letter
}
\examples{
lettergen()
}
\keyword{letter}
